Added comment to parser, refactored python code into one file to avoid circular imports when overriding tensor operators. Added intrinsics for index expressions to python
