+++ python_bindings/pytaco/pytensor/__init__.py
@@ -1,3 +1,2 @@
 from .taco_tensor import *
 from .tensorIO import *
-from .funcs import *
